Create common base for various implementations ServerChannel Actors
